requirements.txt
fastapi
openai
pydantic
uvicorn
python-dotenv